+++ js/background.js
@@ -1,26 +1,44 @@
 const xhttp = new XMLHttpRequest();
+const x2js = new X2JS();
 
 const saleIcon = '../img/notification/icons8-cash-in-hand-96.png',
   refundIcon = '../img/notification/icons8-refund-96.png',
-  chargebackIcon = '../img/notification/icons8-fraud-96.png';
+  chargebackIcon = '../img/notification/icons8-fraud-96.png',
+  reviewIcon = '../img/notification/icons8-very-popular-topic-96.png';
 
-const frequency = { 0: 0, 1: 5, 2: 10, 3: 15, 4: 30, 5: 60 };
+const frequency = {
+  0: 0,
+  1: 5,
+  2: 10,
+  3: 15,
+  4: 30,
+  5: 60
+};
 
 let intervalID,
   interval = 0,
   pubID,
+  pubName,
+  apiKey,
   currentPeriod,
   OpenAsPopupState,
   curWindow,
   lastLink;
 
-function salesLink() {
-  return `https://publisher.assetstore.unity3d.com/api/publisher-info/sales/${pubID}/${currentPeriod}.json`;
+function Links() {
+  return {
+    sales: `https://publisher.assetstore.unity3d.com/api/publisher-info/sales/${pubID}/${currentPeriod}.json`,
+    reviews: `https://publisher.assetstore.unity3d.com/feed/${pubName}/${apiKey}/activity.rss`
+  }
 }
 
+/* start-chrome */
 OpenAsPopupState = localStorage[openAsPopupKey] == null ? true : parseBool(localStorage[openAsPopupKey]);
+/* end-chrome */
 
 StartChecker();
+
+/* start-chrome */
 SetupBrowserAction();
 
 function SetupBrowserAction() {
@@ -73,44 +91,55 @@ function OpenAs(data) {
   localStorage[openAsPopupKey] = OpenAsPopupState;
   SetupBrowserAction();
 }
+/* end-chrome */
 
 chrome.runtime.onMessage.addListener(request => {
-  if (request.command === 'restart') RestartChecker();
-  if (request[updateFrequencyKey] != null) UpdateInterval(request[updateFrequencyKey]);
+  if (request.command === 'restart')
+    RestartChecker();
+  if (request[updateFrequencyKey] != null)
+    UpdateInterval(request[updateFrequencyKey]);
 });
 
 function StartChecker() {
-  chrome.storage.local.get([pubIDKey, currentMonthsKey, updateFrequencyKey], results => {
-    if (results == null || results[pubIDKey] == null || results[currentMonthsKey] == null) return;
+  chrome.storage.local.get([pubIDKey, pubNameKey, apiKeyKey, currentMonthsKey, updateFrequencyKey], results => {
+    if (results == null || results[pubIDKey] == null || results[currentMonthsKey] == null)
+      return;
 
     pubID = results[pubIDKey];
+    pubName = results[pubNameKey];
+    apiKey = results[apiKeyKey];
     currentPeriod = Number(results[currentMonthsKey].replace('-', ''));
     UpdateInterval(results[updateFrequencyKey], true);
   });
 }
 
 function RestartChecker() {
-  if (intervalID != null) clearInterval(intervalID);
+  if (intervalID != null)
+    clearInterval(intervalID);
 
   StartChecker();
 }
 
 function StopChecker() {
-  if (intervalID != null) clearInterval(intervalID);
+  if (intervalID != null)
+    clearInterval(intervalID);
 }
 
 function UpdateInterval(value, force = false) {
-  if (intervalID != null) clearInterval(intervalID);
+  if (intervalID != null)
+    clearInterval(intervalID);
 
-  if (value == null) return;
+  if (value == null)
+    return;
 
   interval = frequency[Number(value)] * 60 * 1000;
-  intervalID = setInterval(DataRequest, interval);
+  intervalID = setInterval(FetchSalesData, interval);
 
-  if (force) DataRequest();
+  if (force)
+    FetchSalesData();
 }
 
-function DataRequest() {
+function FetchSalesData() {
   if (pubID == null || currentPeriod == null) {
     RestartChecker();
     return;
@@ -122,7 +151,7 @@ function DataRequest() {
   }
 
   xhttp.onreadystatechange = HandleStateChange;
-  xhttp.open('GET', salesLink(), true);
+  xhttp.open('GET', Links().sales, true);
   xhttp.timeout = 15000;
   xhttp.send();
 }
@@ -156,27 +185,78 @@ function HandleStateChange() {
       if (previousData == null) {
         localStorage[salesInfoKey] = JSON.stringify(newData);
       } else {
+        localStorage[salesInfoKey] = JSON.stringify(newData);
+
         previousData = JSON.parse(previousData);
 
         if (newData.Qty > previousData.Qty) {
           let diff = newData.Qty - previousData.Qty;
           SendNotification(`You have ${diff > 1 ? diff + ' new sales.' : 'a new sale.'}`, saleIcon);
+          /* start-firefox */
+          return;
+          /* end-firefox */
         }
 
         if (newData.Refunds > previousData.Refunds) {
           let diff = newData.Refunds - previousData.Refunds;
           SendNotification(`You have ${diff > 1 ? diff + ' new refunds.' : 'a new refund.'}`, refundIcon);
+          /* start-firefox */
+          return;
+          /* end-firefox */
         }
 
         if (newData.Chargebacks > previousData.Chargebacks) {
           let diff = newData.Chargebacks - previousData.Chargebacks;
           SendNotification(`You have ${diff > 1 ? diff + ' new chargebacks.' : 'a new chargeback.'}`, chargebackIcon);
         }
-
-        localStorage.setItem(salesInfoKey, JSON.stringify(newData));
       }
-    } else {
-      console.log('Error');
+
+      FetchReviewData();
+    }
+  }
+}
+
+function FetchReviewData() {
+  if (apiKey == null || pubName == null)
+    return;
+
+  xhttp.onreadystatechange = HandleReviewStateChange;
+  xhttp.open('GET', Links().reviews, true);
+  xhttp.setRequestHeader('cache-control', 'max-age=0, no-store');
+  xhttp.timeout = 15000;
+  xhttp.send();
+}
+
+function HandleReviewStateChange() {
+  if (xhttp.readyState === 4) {
+    if (xhttp.status === 200) {
+      let reviewsData = x2js.xml_str2json(xhttp.responseText);
+
+      if (reviewsData.rss.channel.item == null)
+        return;
+
+      let totalReviews = reviewsData.rss.channel.item.length;
+      let userReviews = 0;
+
+      for (let x = 0; x < totalReviews; x++) {
+        let title = reviewsData.rss.channel.item[x].title.replace(/\s+/g, ' ');
+
+        if (title.toLowerCase().includes('reply to review'))
+          continue;
+
+        userReviews++;
+      }
+
+      let previousCount = localStorage[reviewCountKey];
+
+      if (previousCount == null) {
+        localStorage[reviewCountKey] = userReviews;
+      } else if (userReviews > Number(previousCount)) {
+        localStorage[reviewCountKey] = userReviews;
+
+        let diff = userReviews - Number(previousCount);
+        SendNotification(`You have ${diff > 1 ? diff + ' new reviews.' : 'a new review.'}`, reviewIcon);
+      }
     }
   }
 }
@@ -190,17 +270,36 @@ function SendNotification(body, icon) {
   });
 }
 
-chrome.contextMenus.create({
-  title: 'Verify Invoice',
-  contexts: ['selection'],
-  onclick: VerifyInvoice
-});
+chrome.runtime.onInstalled.addListener(() => {
+  chrome.contextMenus.create({
+    id: 'verify-invoice',
+    title: 'Verify Invoice',
+    contexts: ['selection'],
+    visible: true,
+    onclick: VerifyInvoice
+  });
+})
 
 function VerifyInvoice(data) {
   let selectedText = data.selectionText.replace(/[^\w\s]/gi, '').replace(/ /g, '');
+  /* start-chrome */
   OpenWindow(`../index.html?inv=${selectedText}&resize`);
+  /* end-chrome */
+
+  /* start-firefox */
+  browser.browserAction.setPopup({
+    popup: `../index.html?inv=${selectedText}`
+  });
+
+  browser.browserAction.openPopup();
+
+  browser.browserAction.setPopup({
+    popup: "../index.html"
+  });
+  /* end-firefox */
 }
 
+/* start-chrome */
 function OpenWindow(link) {
   let width, height, top, left;
 
@@ -211,8 +310,7 @@ function OpenWindow(link) {
   left = localStorage[popupWindowLeftPosKey] != null ? Number(localStorage[popupWindowLeftPosKey]) : null;
 
   if (curWindow == null) {
-    chrome.windows.create(
-      {
+    chrome.windows.create({
         url: link,
         type: 'popup',
         width: width,
@@ -242,11 +340,15 @@ function OpenWindow(link) {
 }
 
 chrome.windows.onRemoved.addListener(windowId => {
-  if (curWindow == null) return;
+  if (curWindow == null)
+    return;
 
-  if (windowId == curWindow.id) curWindow = null;
+  if (windowId == curWindow.id)
+    curWindow = null;
 });
+
 
 function parseBool(b) {
   return !/^(false|0)$/i.test(b) && !!b;
 }
+/* end-chrome */